Added transform handling and added more tests to the math library

+++ eclipse_workspace/engine-core/src/main/java/com/vine/math/VectorOperations.java
@@ -16,7 +16,7 @@ public interface VectorOperations<V> {
      * Returns the dot product of this Vector2f and the given. Callers guarantee
      * that the given vector is a valid object.
      */
-    double dot(final V vector);
+    float dot(final V vector);
 
     /**
      * Multiplies the elements of this vector with the given scale float value.
